Select language modifications
Now dialog to change language is properly working and updates intantly.

When the user first installs the app, it will set the user config to en or es if the phone uses that language, if not, it will use english. If user select a language, it will be used that instead.

Removed language from user config, now is completelly separated and it will keep the same language to all accounts in the same device.

Login page added text open the change language dialog.